fix: extended the genbank parser to include all features, corrected FPKM total number of reads.
This commit introduces a few fixes. 1) The total number of reads considered for the FPKM normalization is now taken from the number of reads in the reads BED file, n_reads_bed, *without removing the rRNA reads*. FPKM normalization should be relative to the total number of reads that map to the genome, and should not depend on the annotation of the features that may vary. Before, we just summed up the nb of counted fragments for all the features in the dataframe: then, the total nb of reads will depend on the exact list of features, and might be overestimated if features are overlapping. 2) The genbank parser now takes into account all features in the genbank file, e.g. both gene and CDS. Before it only was keeping (CDS)|(tRNA)|(ncRNA)|(sRNA). Because there will be duplicated features for the same coordinates, e.g. gene and CDS with the same id, and the BED format does not describe the feature *type*, we simply create all features and then remove duplicates in the BED file.